requirements.txt
streamlit
langchain
langgraph
langchain-huggingface
huggingface_hub
langchain-community
pypdf
beautifulsoup4
wikipedia
python-dotenv
tiktoken
faiss-cpu
sentence-transformers
